requirements.txt
et_xmlfile==2.0.0
lxml==5.3.0
openpyxl==3.1.5
python-docx==1.1.2
typing_extensions==4.12.2
